Simplify Basisfunction compilation
