+++ src/sources.cmake
@@ -204,6 +204,7 @@ target_sources(preciceCore
     src/mapping/config/MappingConfiguration.cpp
     src/mapping/config/MappingConfiguration.hpp
     src/mapping/config/MappingConfigurationTypes.hpp
+    src/mapping/impl/BasisFunctions.cpp
     src/mapping/impl/BasisFunctions.hpp
     src/mapping/impl/CreateClustering.hpp
     src/mapping/impl/SphericalVertexCluster.hpp
